+++ internal/chain/util_test.go
@@ -35,7 +35,7 @@ func TestIsValidAddress(t *testing.T) {
 func TestEtherToWei(t *testing.T) {
 	tests := []struct {
 		name   string
-		amount int64
+		amount float64
 		want   *big.Int
 	}{
 		{name: "1ether", amount: 1, want: new(big.Int).Exp(big.NewInt(10), big.NewInt(18), nil)},
